feat: Partially add Kubernetes list library
This commit adds the isSorted, indexOf and lastIndexOf methods on
lists. These functions are modelled after the Kubernetes list
library/extension.
See https://kubernetes.io/docs/reference/using-api/cel/#kubernetes-list-library
