+++ cel/src/functions/k8s/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(feature = "k8s-list")]
+pub mod list;
